COOK-143 - implement uc: show recipe of the month: extracted showRecipeOfTheMonth method to GlobalControllerAdvice
